+++ packages/frontend/src/game/menu/utility/drawMenuContainer.ts
@@ -5,11 +5,11 @@ export const drawMenuContainer = () => {
     const menuContainer = new Rectangle();
     menuContainer.width = startScreenUIDimensions.MENU_CONTAINER.WIDTH;
     menuContainer.height = startScreenUIDimensions.MENU_CONTAINER.HEIGHT;
-    
-    menuContainer.background = startScreenConfig.UI_CONFIG.MENU_CONTAINER.BACKGROUND;
-    menuContainer.cornerRadius = startScreenConfig.UI_CONFIG.MENU_CONTAINER.CORNER_RADIUS;
     menuContainer.thickness = startScreenConfig.UI_CONFIG.MENU_CONTAINER.BORDER_THICKNESS;
+    menuContainer.cornerRadius = startScreenConfig.UI_CONFIG.MENU_CONTAINER.CORNER_RADIUS;
+    menuContainer.background = startScreenConfig.UI_CONFIG.MENU_CONTAINER.BACKGROUND;
     menuContainer.color = startScreenConfig.UI_CONFIG.MENU_CONTAINER.COLOR;
+    
     menuContainer.shadowColor = startScreenConfig.UI_CONFIG.MENU_CONTAINER.SHADOW_COLOR;
     menuContainer.shadowBlur = startScreenConfig.UI_CONFIG.MENU_CONTAINER.SHADOW_BLUR;
     menuContainer.shadowOffsetX = startScreenConfig.UI_CONFIG.MENU_CONTAINER.SHADOW_OFFSET_X;
